@@ -1,13 +1,14 @@
 apply plugin: 'com.android.application'
 
 android {
-    compileSdkVersion 23
-    buildToolsVersion "23.0.2"
+    namespace 'jakhar.aseem.diva'
+    compileSdk 33
+
 
     defaultConfig {
         applicationId "jakhar.aseem.diva"
-        minSdkVersion 15
-        targetSdkVersion 23
+        minSdk 15
+        targetSdk 33
         versionCode 1
         versionName "1.0"
     }
@@ -17,16 +18,29 @@ android {
             proguardFiles getDefaultProguardFile('proguard-android.txt'), 'proguard-rules.pro'
         }
     }
+    compileOptions {
+        sourceCompatibility JavaVersion.VERSION_1_8
+        targetCompatibility JavaVersion.VERSION_1_8
+    }
     sourceSets {
         main {
             jni.srcDirs = []
         }
     }
+    externalNativeBuild {
+        cmake {
+            version "3.22.1"
+        }
+    }
+    ndkVersion '25.2.9519653'
+    buildToolsVersion '33.0.0'
 }
 
 dependencies {
-    compile fileTree(dir: 'libs', include: ['*.jar'])
-    testCompile 'junit:junit:4.12'
-    compile 'com.android.support:appcompat-v7:23.1.0'
-    compile 'com.android.support:design:23.1.0'
+    implementation fileTree(dir: 'libs', include: ['*.jar'])
+    implementation 'androidx.appcompat:appcompat:1.6.1'
+    implementation 'com.android.support:design:28.0.1'
+    testImplementation 'junit:junit:4.13.2'
+    androidTestImplementation 'androidx.test.ext:junit:1.1.5'
+    androidTestImplementation 'androidx.test.espresso:espresso-core:3.5.1'
 }
