@@ -31,8 +31,8 @@
  */
 package jakhar.aseem.diva;
 
-import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
+import androidx.appcompat.app.AppCompatActivity;
 import android.widget.TextView;
 
 public class APICredsActivity extends AppCompatActivity {
